+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
-## [0.1.0] - 2023-09-19
+## [0.1.0] - 2023-09-20
 
 - Initial release
+  - Add `Datadog::Logging::Logger#http_request` and `Datadog::Logging::Logger::UserFormatter`
+  - Add `Datadog::Logging::Formatter::JsonFormatter`
+  - Add `Datadog::Logging::Middleware::LogExceptions`
+  - Add `Datadog::Logging::Railtie`